# docs/swagger.yaml
basePath: /
definitions:
  server.CreateTaskRequest:
    description: Request body for creating a new task
    properties:
      areaName:
        description: Area name to assign the task to
        example: Personal
        type: string
      description:
        description: Optional description
        example: Get milk, eggs, and bread
        type: string
      dueDate:
        description: Due date for the task (RFC3339 format)
        example: "2025-01-20T00:00:00Z"
        type: string
      plannedDate:
        description: Planned date for the task (RFC3339 format)
        example: "2025-01-15T00:00:00Z"
        type: string
      projectName:
        description: Project name to assign the task to
        example: Shopping
        type: string
      recurEnd:
        description: Recurrence end date (RFC3339 format)
        example: "2025-12-31T00:00:00Z"
        type: string
      recurRule:
        description: Recurrence rule as JSON
        example: '{"interval":1,"unit":"week"}'
        type: string
      recurType:
        description: 'Recurrence type: "fixed" or "relative"'
        example: fixed
        type: string
      someday:
        description: Set to true to create task in someday state
        example: false
        type: boolean
      tags:
        description: Tags to assign to the task
        example:
        - urgent
        - important
        items:
          type: string
        type: array
      title:
        description: Title of the task (required)
        example: Buy groceries
        type: string
    type: object
  server.ErrorResponse:
    description: Error response from the API
    properties:
      error:
        description: Error message
        example: Task not found
        type: string
    type: object
  server.SetRecurrenceRequest:
    description: Request body for setting recurrence on a task
    properties:
      recurEnd:
        description: Optional recurrence end date (RFC3339 format)
        example: "2025-12-31T00:00:00Z"
        type: string
      recurRule:
        description: Recurrence rule as JSON
        example: '{"interval":1,"unit":"week"}'
        type: string
      recurType:
        description: 'Recurrence type: "fixed" or "relative"'
        example: fixed
        type: string
    type: object
  server.UpdateTaskRequest:
    description: Request body for updating an existing task (all fields optional)
    properties:
      areaName:
        description: Area name to assign the task to
        example: Home
        type: string
      clearArea:
        description: Set to true to unassign from area
        example: false
        type: boolean
      clearDescription:
        description: Set to true to clear the description
        example: false
        type: boolean
      clearDueDate:
        description: Set to true to clear the due date
        example: false
        type: boolean
      clearPlannedDate:
        description: Set to true to clear the planned date
        example: false
        type: boolean
      clearProject:
        description: Set to true to unassign from project
        example: false
        type: boolean
      description:
        description: New description for the task
        example: Updated description
        type: string
      dueDate:
        description: New due date (RFC3339 format)
        example: "2025-01-20T00:00:00Z"
        type: string
      plannedDate:
        description: New planned date (RFC3339 format)
        example: "2025-01-15T00:00:00Z"
        type: string
      projectName:
        description: Project name to assign the task to
        example: Work
        type: string
      state:
        description: 'Task state: "active" or "someday"'
        example: active
        type: string
      tags:
        description: Replace all tags with these
        example:
        - work
        - priority
        items:
          type: string
        type: array
      title:
        description: New title for the task
        example: Updated task title
        type: string
    type: object
  task.CompleteResult:
    properties:
      completed:
        $ref: '#/definitions/task.Task'
      nextTask:
        allOf:
        - $ref: '#/definitions/task.Task'
        description: non-nil if a recurring task was regenerated
    type: object
  task.State:
    enum:
    - active
    - someday
    type: string
    x-enum-varnames:
    - StateActive
    - StateSomeday
  task.Status:
    enum:
    - todo
    - done
    type: string
    x-enum-varnames:
    - StatusTodo
    - StatusDone
  task.Task:
    properties:
      areaId:
        type: integer
      areaName:
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: integer
      parentId:
        type: integer
      parentName:
        description: Display fields (populated by queries with JOINs, not persisted)
        type: string
      plannedDate:
        type: string
      recurEnd:
        description: optional end date
        type: string
      recurParentId:
        description: links to original recurring task
        type: integer
      recurPaused:
        description: true = paused
        type: boolean
      recurRule:
        description: 'JSON rule: {"interval":1,"unit":"week",...}'
        type: string
      recurType:
        description: Recurrence fields
        type: string
      state:
        $ref: '#/definitions/task.State'
      status:
        $ref: '#/definitions/task.Status'
      tags:
        description: Tags
        items:
          type: string
        type: array
      taskType:
        $ref: '#/definitions/task.TaskType'
      title:
        type: string
      uuid:
        type: string
    type: object
  task.TaskType:
    enum:
    - task
    - project
    type: string
    x-enum-varnames:
    - TaskTypeTask
    - TaskTypeProject
host: localhost:8080
info:
  contact: {}
  description: Task management sync server with CRUD operations
  title: TT Sync Server API
  version: "1.0"
paths:
  /api/v1/tasks:
    get:
      description: List tasks with optional filters
      parameters:
      - description: Filter by project name
        in: query
        name: project
        type: string
      - description: Filter by area name
        in: query
        name: area
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Search in task title
        in: query
        name: search
        type: string
      - description: 'Filter by schedule: today, upcoming, anytime, inbox, someday'
        in: query
        name: schedule
        type: string
      - description: 'Sort order: field:dir,field:dir (e.g. due:asc,title:desc)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the given parameters
      parameters:
      - description: Task creation parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /api/v1/tasks/{uuid}:
    delete:
      description: Delete a task by its UUID
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Retrieve a single task by its UUID
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a task by UUID
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update a task's properties (partial update)
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
  /api/v1/tasks/{uuid}/complete:
    post:
      description: Mark a task as completed
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.CompleteResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete a task
      tags:
      - tasks
  /api/v1/tasks/{uuid}/recurrence:
    patch:
      consumes:
      - application/json
      description: Configure recurrence settings for a task
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Recurrence settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.SetRecurrenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set task recurrence
      tags:
      - tasks
  /api/v1/tasks/{uuid}/recurrence/pause:
    post:
      description: Pause the recurrence on a recurring task
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Pause task recurrence
      tags:
      - tasks
  /api/v1/tasks/{uuid}/recurrence/resume:
    post:
      description: Resume a paused recurrence on a task
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resume task recurrence
      tags:
      - tasks
  /api/v1/tasks/{uuid}/uncomplete:
    post:
      description: Mark a task as not completed
      parameters:
      - description: Task UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Uncomplete a task
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" where {token} is your API key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"